Move serialization outside of cache lock
